Add space home layout and assistant prompt
- Split the space Home view into recent objects, welcome, and prompt regions
- Add starter suggestions and a pinned assistant prompt on the Home page
- Move the virtual Home node into plugin-space and wire toolbar actions
