+++ packages/plugins/plugin-assistant/src/containers/SpaceHomePrompt/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export { SpaceHomePrompt as default } from './SpaceHomePrompt';
