+++ packages/plugins/plugin-assistant/src/containers/SpaceHomeSuggestions/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export { SpaceHomeSuggestions as default } from './SpaceHomeSuggestions';
