+++ packages/plugins/plugin-assistant/src/containers/index.ts
@@ -13,5 +13,7 @@ export const ChatDialog: ComponentType<any> = lazy(() => import('./ChatDialog'))
 export const PlanArticle: ComponentType<any> = lazy(() => import('./PlanArticle'));
 export const RoutineArticle: ComponentType<any> = lazy(() => import('./RoutineArticle'));
 export const RoutineList: ComponentType<any> = lazy(() => import('./RoutineList'));
+export const SpaceHomePrompt: ComponentType<any> = lazy(() => import('./SpaceHomePrompt'));
+export const SpaceHomeSuggestions: ComponentType<any> = lazy(() => import('./SpaceHomeSuggestions'));
 export const TracePanel: ComponentType<any> = lazy(() => import('./TracePanel'));
 export const TriggerStatus: ComponentType<any> = lazy(() => import('./TriggerStatus'));
