+++ packages/plugins/plugin-space/src/containers/SpaceHomeRecent/SpaceHomeRecent.tsx
@@ -1,0 +1,113 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+import * as Option from 'effect/Option';
+import React, { useCallback, useMemo } from 'react';
+
+import { useCapabilities, useOperationInvoker } from '@dxos/app-framework/ui';
+import { AppCapabilities, LayoutOperation, getObjectPathFromObject } from '@dxos/app-toolkit';
+import { Collection, Filter, Obj, Order, Query, Type } from '@dxos/echo';
+import { HiddenAnnotation, getTypeAnnotation } from '@dxos/echo/Annotation';
+import { Kind as EntityKind } from '@dxos/echo/Entity';
+import { type Space, useQuery } from '@dxos/react-client/echo';
+import { Card, Icon, toLocalizedString, useTranslation } from '@dxos/react-ui';
+import { Masonry } from '@dxos/react-ui-masonry';
+import { getStyles } from '@dxos/ui-theme';
+
+import { meta } from '#meta';
+
+/** Number of recently-modified objects to surface as cards. */
+const RECENT_LIMIT = 10;
+
+type SpaceScopedProps = {
+  space?: Space;
+};
+
+/**
+ * Recent-objects region for the Home article. Queries the most-recently-modified objects of
+ * registered, non-hidden, non-relation, non-collection types and renders them as a Masonry of
+ * tiles. Renders nothing (no heading) when the space has no recent objects — the starter-prompt
+ * contributor (plugin-assistant) fills the empty state instead.
+ */
+export const SpaceHomeRecent = ({ space }: SpaceScopedProps) => {
+  const { t } = useTranslation(meta.id);
+
+  const schemas = useCapabilities(AppCapabilities.Schema);
+  const filter = useMemo(() => {
+    const collectionTypename = Type.getTypename(Collection.Collection);
+    const types = schemas
+      .flat()
+      .filter(Type.isType)
+      .filter((type) => getTypeAnnotation(Type.getSchema(type))?.kind !== EntityKind.Relation)
+      .filter((type) => !HiddenAnnotation.get(Type.getSchema(type)).pipe(Option.getOrElse(() => false)))
+      .filter((type) => Type.getTypename(type) !== collectionTypename);
+    return types.length > 0 ? Filter.or(...types.map((type) => Filter.type(type))) : undefined;
+  }, [schemas]);
+
+  const query = useMemo(
+    () =>
+      Query.select(filter ?? Filter.everything())
+        .orderBy(Order.updated('desc'))
+        .limit(RECENT_LIMIT),
+    [filter],
+  );
+  const recent = useQuery(filter && space ? space.db : undefined, query);
+
+  const RecentTile = useMemo(() => {
+    const Tile = ({ data }: { data: Obj.Unknown; index: number }) => <RecentObjectTile space={space} object={data} />;
+    Tile.displayName = 'RecentObjectTile';
+    return Tile;
+  }, [space]);
+
+  if (recent.length === 0) {
+    return null;
+  }
+
+  return (
+    <>
+      <h2 className='text-sm font-medium text-description'>{t('space-home.recent.heading')}</h2>
+      <Masonry.Root Tile={RecentTile}>
+        <Masonry.Content padding={false}>
+          <Masonry.Viewport classNames='py-2' items={recent} getId={(object) => object.id} />
+        </Masonry.Content>
+      </Masonry.Root>
+    </>
+  );
+};
+
+type RecentObjectTileProps = {
+  space?: Space;
+  object: Obj.Unknown;
+};
+
+const RecentObjectTile = ({ space, object }: RecentObjectTileProps) => {
+  const { invokePromise } = useOperationInvoker();
+  const { t } = useTranslation(meta.id);
+  const typename = Obj.getTypename(object);
+  const label = toLocalizedString(
+    Obj.getLabel(object) ?? (typename ? ['object-name.placeholder', { ns: typename, defaultValue: 'New item' }] : ''),
+    t,
+  );
+  const iconAnnotation = Obj.getIcon(object);
+  const icon = iconAnnotation?.icon ?? 'ph--circle-dashed--regular';
+  const iconStyles = iconAnnotation?.hue ? getStyles(iconAnnotation.hue) : undefined;
+
+  const handleClick = useCallback(() => {
+    if (!space) {
+      return;
+    }
+    void invokePromise(LayoutOperation.Open, { subject: [getObjectPathFromObject(object)] });
+  }, [invokePromise, object, space]);
+
+  return (
+    <Card.Root role='button' classNames='cursor-pointer' onClick={handleClick}>
+      <Card.Header>
+        <Card.Block>
+          <Icon icon={icon} classNames={iconStyles?.text} />
+        </Card.Block>
+        <Card.Title>{label}</Card.Title>
+      </Card.Header>
+    </Card.Root>
+  );
+};
