+++ packages/plugins/plugin-space/src/containers/SpaceHomeRecent/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export { SpaceHomeRecent as default } from './SpaceHomeRecent';
