+++ packages/plugins/plugin-support/src/containers/SpaceHomeWelcome/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export { SpaceHomeWelcome as default } from './SpaceHomeWelcome';
