+++ packages/plugins/plugin-support/src/containers/index.ts
@@ -10,6 +10,6 @@ export const HelpMenu: ComponentType<any> = lazy(() => import('./HelpMenu'));
 export const ShortcutsDialogContent: ComponentType<any> = lazy(() => import('./ShortcutsDialogContent'));
 export const ShortcutsHints: ComponentType<any> = lazy(() => import('./ShortcutsHints'));
 export const ShortcutsList: ComponentType<any> = lazy(() => import('./ShortcutsList'));
-export const SpaceHomeArticle: ComponentType<any> = lazy(() => import('./SpaceHomeArticle'));
+export const SpaceHomeWelcome: ComponentType<any> = lazy(() => import('./SpaceHomeWelcome'));
 export const SupportArticle: ComponentType<any> = lazy(() => import('./SupportArticle'));
 export const SupportCompanion: ComponentType<any> = lazy(() => import('./SupportCompanion'));
