+++ packages/plugins/plugin-support/src/operations/index.ts
@@ -7,6 +7,7 @@ import { OperationHandlerSet } from '@dxos/compute';
 export const SupportOperationHandlerSet = OperationHandlerSet.lazy(
   () => import('./capture-feedback'),
   () => import('./create-ticket'),
+  () => import('./hide-welcome'),
   () => import('./mark-in-progress'),
   () => import('./on-create-space'),
   () => import('./resolve-ticket'),
