+++ packages/plugins/plugin-support/src/translations.ts
@@ -69,15 +69,9 @@ export const translations = [
         'open-help-tour.message': 'Show welcome tour',
         'open-shortcuts.label': 'Show shortcuts',
         'shortcuts-dialog.title': 'Shortcuts',
-        // Per-space Home node + Welcome content.
-        'space-home-node.label': 'Home',
-        'space-home.recent.heading': 'Recent',
-        'space-home.suggestions.heading': 'Get started',
-        'space-home.suggestion-draft-doc.label': 'Draft a new document',
-        'space-home.suggestion-ideas.label': 'Suggest some ideas to work on',
-        'space-home.suggestion-data-type.label': 'Create a new data type',
-        'space-home.prompt.placeholder': 'Ask the assistant anything…',
-        'space-home.empty.label': 'Create an object to get started.',
+        // Per-space Home Welcome content. The node label, recent heading, suggestions, and prompt
+        // strings live in plugin-space/plugin-assistant; the toolbar actions are contributed as
+        // graph actions (see app-graph-builder) and use the labels below.
         'welcome.title': 'Welcome to Composer',
         'welcome.description':
           'This is your personal space. Open a ticket from the Support assistant, or start the guided tour to learn how to get started.',
