+++ packages/plugins/plugin-support/tsconfig.json
@@ -63,6 +63,9 @@
       "path": "../../sdk/app-toolkit"
     },
     {
+      "path": "../../sdk/client"
+    },
+    {
       "path": "../../sdk/client-protocol"
     },
     {
