+++ packages/sdk/app-toolkit/moon.yml
@@ -11,6 +11,7 @@ tasks:
       - '--entryPoint=src/index.ts'
       - '--entryPoint=src/activation-events.ts'
       - '--entryPoint=src/query.ts'
+      - '--entryPoint=src/space-home.ts'
       - '--entryPoint=src/ui/index.ts'
   test:
     inputs:
