+++ packages/sdk/app-toolkit/src/index.ts
@@ -21,6 +21,7 @@ export * from './paths';
 export * from './plugin';
 export * from './native-passkey';
 export * from './personal-space';
+export * from './space-home';
 export * from './type-options';
 export * from './type-section-extension';
 export * from './type-section-resolver';
